rafactor: rename custom components directory
